[Refactor] index.js로 경로 변경 #39
