+++ src/components/Header/Header.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Nav from '@/components/Nav/Nav';
+import Nav from '@/components/index';
 import Logo from '@/assets/Logo.svg';
 import styled from 'styled-components/macro';
